+++ src/cpp/multisolver_interface/environment.cc
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "environment.h"
+#include "multisolver_interface/environment.h"
 
 #include <filesystem>
 #include <iostream>
