+++ tests/cpp/sensitivity/SensitivityStudyTest.cpp
@@ -1,8 +1,8 @@
 #include "SensitivityFileLogger.h"
 #include "SensitivityStudy.h"
-#include "environment.h"
 #include "gtest/gtest.h"
 #include "multisolver_interface/SolverFactory.h"
+#include "multisolver_interface/environment.h"
 
 class SensitivityStudyTest : public ::testing::Test {
  public:
